move tasks out of pipeline module
