+++ include/dftracer/utils/core/tasks/tasks.h
@@ -1,0 +1,10 @@
+#ifndef DFTRACER_UTILS_CORE_TASKS_TASKS_H
+#define DFTRACER_UTILS_CORE_TASKS_TASKS_H
+
+#include <dftracer/utils/core/tasks/function_task.h>
+#include <dftracer/utils/core/tasks/task.h>
+#include <dftracer/utils/core/tasks/task_context.h>
+#include <dftracer/utils/core/tasks/task_result.h>
+#include <dftracer/utils/core/tasks/task_tag.h>
+
+#endif  // DFTRACER_UTILS_CORE_TASKS_TASKS_H
