+++ include/dftracer/utils/core/tasks/typed_task.h
@@ -1,8 +1,8 @@
-#ifndef DFTRACER_UTILS_CORE_PIPELINE_TASKS_TYPED_TASK_H
-#define DFTRACER_UTILS_CORE_PIPELINE_TASKS_TYPED_TASK_H
+#ifndef DFTRACER_UTILS_CORE_TASKS_TYPED_TASK_H
+#define DFTRACER_UTILS_CORE_TASKS_TYPED_TASK_H
 
 #include <dftracer/utils/core/common/logging.h>
-#include <dftracer/utils/core/pipeline/tasks/task.h>
+#include <dftracer/utils/core/tasks/task.h>
 
 #include <typeindex>
 
@@ -41,4 +41,4 @@ class TypedTask : public Task {
 
 }  // namespace dftracer::utils
 
-#endif  // DFTRACER_UTILS_CORE_PIPELINE_TASKS_TYPED_TASK_H
+#endif  // DFTRACER_UTILS_CORE_TASKS_TYPED_TASK_H
