+++ src/dftracer/utils/core/pipeline/executors/scheduler/sequential_scheduler.cpp
@@ -3,7 +3,7 @@
 #include <dftracer/utils/core/pipeline/executors/executor_context.h>
 #include <dftracer/utils/core/pipeline/executors/scheduler/sequential_scheduler.h>
 #include <dftracer/utils/core/pipeline/pipeline.h>
-#include <dftracer/utils/core/pipeline/tasks/task_context.h>
+#include <dftracer/utils/core/tasks/task_context.h>
 
 #include <any>
 #include <unordered_map>
